+++ history.md
@@ -2,6 +2,7 @@
 
 - Placeholder and update value support for `InputBox` and `QuickPick`.
 - `triggerOnly` option property for vim completion source.
+- Export `getExtensionById` from `extensions` module.
 
 # 2022-11-26